Rewrite API
Several of the API methods now allow users to specify the necessary
column names for each procedure, but their default methods are aligned
with the original naming convention. This allows users to have a wide
variety of versatility in applying these methods while also just being
able to use more concise versions of these methods when following the
standard naming conventions.

Also adds an inplace parameter to most of the methods that manipulate
the data so that users can choose between getting a copy or modifying
the data inplace.